refactor: local-storage 관련 로직들 유틸함수로 분리
